requirements.txt
streamlit
pdfplumber
googletrans==4.0.0-rc1
fpdf
python-docx